UserFunction: Missing cast added.
git-svn-id: https://svn.berlios.de/svnroot/repos/escript/branches/0.6.5.experimental@443 5693b1fc-3b75-4070-9b6b-3ce3692a40d5

+++ EScript/Objects/Callables/UserFunction.cpp
@@ -37,7 +37,7 @@ void UserFunction::init(EScript::Namespace & globals) {
 	ESMF_DECLARE(t,UserFunction,"getMinParamCount",0,0, self->getMinParamCount())
 
 	//! [ESMF] Number UserFunction.getParamCount()
-	ESMF_DECLARE(t,UserFunction,"getParamCount",0,0, self->getParamCount())
+	ESMF_DECLARE(t,UserFunction,"getParamCount",0,0, static_cast<uint32_t>(self->getParamCount()))
 
 	//! [ESMF] String UserFunction._asm()
 	ESMF_DECLARE(t,UserFunction,"_asm",0,0, self->getInstructionBlock().toString())
